# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoyaStar Backend API",
        "description": "LMS platform for managing students, teachers, schools, and curriculum.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "List all students",
                "operationId": "0df87646b50d15aea77868bb24a975ec",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grade_id",
                        "in": "query",
                        "description": "Filter by grade ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "description": "Filter by city ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "school_name",
                        "in": "query",
                        "description": "Filter by school name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort column (e.g. name, created_at)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 20)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of students"
                    }
                },
                "security": [
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create student",
                "operationId": "5c9b91c56de266e807f1492095c80fcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "grade_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "school_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created"
                    }
                }
            }
        },
        "/students/{student}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get student by ID",
                "operationId": "b07fc5d36d18f74880bc1a42d1bb8f3c",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student data"
                    }
                },
                "security": [
                    {}
                ]
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update student",
                "operationId": "923393f56a6f512daab6beeb6173d8fc",
                "responses": {
                    "200": {
                        "description": "Student updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete student",
                "operationId": "a3309f52e5c198745d464b0e1712e7b8",
                "responses": {
                    "200": {
                        "description": "Student deleted"
                    }
                }
            }
        },
        "/students/activities": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get authenticated student activities",
                "operationId": "1a1cb159895bdc08f023dcb502543db4",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by activity type (lesson_completed, quiz_completed)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 20)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of activities"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "bearer",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}